requirements.txt
azure-ai-ml>=1.0.0
mlflow>=2.0.0
scikit-learn>=1.0.0
pandas>=1.3.0
PyYAML>=5.4.0
numpy>=1.21.0
faker>=15.0.0